+++ scripts/prepare_build_py35.sh
@@ -8,4 +8,10 @@ sed -i '' -e 's/unicode(/str(/g' $(find src_py35 -type f)
 python setup.py sdist bdist_wheel --python-tag py35
 rm -rf dist_py35/
 mv dist/ dist_py35/
-rm dist_py35/*.tar.gz
+
+if [[ $1 == "--keep-tar-gz" ]]
+then
+  shift
+else
+  rm dist_py35/*.tar.gz
+fi
